+++ CHANGELOG.md
@@ -1,5 +1,44 @@
 # CHANGELOG
 
+## [1.6.50](https://github.com/deepmodeling/dflow/compare/v1.6.49...v1.6.50) (2023-02-16)
+
+
+### Bug Fixes
+
+* big parameter string null bug ([f9925c2](https://github.com/deepmodeling/dflow/commit/f9925c2c6b8de435e0314b6011a8a24335e492cb))
+
+## [1.6.50]
+
+### Bug Fixes
+
+* Support BigParameter default value on local machine
+
+## [1.6.49](https://github.com/deepmodeling/dflow/compare/v1.6.48...v1.6.49) (2023-02-15)
+
+
+### Bug Fixes
+
+* deprecate lebesgue executor ([0a68cb9](https://github.com/deepmodeling/dflow/commit/0a68cb9663d320744c6d2f8ed435bae8e6a020b3))
+* one task not failing the whole group when pool_size used ([0a68cb9](https://github.com/deepmodeling/dflow/commit/0a68cb9663d320744c6d2f8ed435bae8e6a020b3))
+
+## [1.6.48](https://github.com/deepmodeling/dflow/compare/v1.6.47...v1.6.48) (2023-02-14)
+
+
+### Bug Fixes
+
+* direct jsonpickle.dumps for big parameter ([7e1a063](https://github.com/deepmodeling/dflow/commit/7e1a063fbcfdd7e4ddbcaff3e0b9f761f1fc720f))
+* support for argo_len of big parameter ([7e1a063](https://github.com/deepmodeling/dflow/commit/7e1a063fbcfdd7e4ddbcaff3e0b9f761f1fc720f))
+
+## [1.6.47](https://github.com/deepmodeling/dflow/compare/v1.6.46...v1.6.47) (2023-02-13)
+
+
+### Bug Fixes
+
+* add API for creating bohrium job group ID ([772ad4d](https://github.com/deepmodeling/dflow/commit/772ad4ddb0f89e2dde6c0a0a231fa7c944788e6e))
+* process cannot pass dflow configs ([77febb9](https://github.com/deepmodeling/dflow/commit/77febb9c7d618bfd62cac8f8c504e8d707909196))
+* support for expression of loop item ([12c58c9](https://github.com/deepmodeling/dflow/commit/12c58c9e6cc6caff593a3e524f4eb882d492561d))
+* support for slicing big parameter ([772ad4d](https://github.com/deepmodeling/dflow/commit/772ad4ddb0f89e2dde6c0a0a231fa7c944788e6e))
+
 ## [1.6.46](https://github.com/deepmodeling/dflow/compare/v1.6.45...v1.6.46) (2023-02-10)